Update schema to represent non drug types. Also extended Recommendations
Schema is now extended with an ontology for non drug types and their execution. Both the type and its execution are within classes that contain also drug types and their administration. This makes their definitions non-conflictive with the TMR guideline definition  rules and interaction rules. Additionally, the vocabulary has been extended to contain also 'motivation' as an explanation for a recommendation to be or not be recommended. This means that a recommendation has now an extra rdf but it does not need to be added to the recommendation rules since it does not affect the outcome but provides extra information for the user.